Move broken_link to other modules.
